+++ circular-deps.txt
@@ -1,4 +1,4 @@
-Processed 152 files (1s) (16 warnings)
+Processed 152 files (995ms) (16 warnings)
 
 1) Config/Settings.ts > Suggestor/Suggestor.ts
 2) Task/Task.ts > Config/Settings.ts > Suggestor/Suggestor.ts
@@ -9,8 +9,7 @@ Processed 152 files (1s) (16 warnings)
 7) Task/Task.ts > lib/LogTasksHelper.ts
 8) Query/Query.ts > Query/Explain/Explainer.ts
 9) Query/FilterParser.ts > Query/Filter/BooleanField.ts
-10) Statuses/StatusRegistryReport.ts > Statuses/StatusSettingsReport.ts
-11) main.ts > Config/SettingsTab.ts
-12) Obsidian/Cache.ts > Obsidian/TasksEvents.ts
-13) main.ts > Renderer/QueryRenderer.ts
+10) main.ts > Config/SettingsTab.ts
+11) Obsidian/Cache.ts > Obsidian/TasksEvents.ts
+12) main.ts > Renderer/QueryRenderer.ts
