+++ src/Statuses/StatusRegistryReport.ts
@@ -47,14 +47,3 @@ ${settingsTable}
 These are the settings actually used by Tasks.
 ${mermaidText}`;
 }
-
-export function getPrintableSymbol(symbol: string) {
-    // Do not put backticks around an empty symbol, as the two backticks are rendered
-    // by Obsidian as ordinary characters and the meaning is unclear.
-    // Better to just display nothing in this situation.
-    if (symbol === '') {
-        return symbol;
-    }
-    const result = symbol !== ' ' ? symbol : 'space';
-    return '`' + result + '`';
-}
